requirements.txt
instaloader==4.10.3
schedule==1.2.0
requests==2.31.0
python-dotenv==1.0.0
notion-client==2.2.1